Refactored tests / DRY.
