+++ backend/src/test/java/com/github/boltzmann/biokiste/backend/controller/CrudWithLogInTest.java
@@ -1,0 +1,65 @@
+package com.github.boltzmann.biokiste.backend.controller;
+
+import com.github.boltzmann.biokiste.backend.repository.AppUserDetailsRepo;
+import com.github.boltzmann.biokiste.backend.repository.BoxRepository;
+import com.github.boltzmann.biokiste.backend.security.model.AppUser;
+import com.github.boltzmann.biokiste.backend.security.repository.AppUserLoginRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class CrudWithLogInTest {
+    @Value("${piphi.biokisteapp.jwt.secret}")
+    private String jwtSecret;
+
+    @Autowired
+    PasswordEncoder passwordEncoder;
+
+    @Autowired
+    AppUserLoginRepository appUserLoginRepository;
+
+    @Autowired
+    WebTestClient webTestClient;
+
+    @Autowired
+    BoxRepository boxRepository;
+
+    @Autowired
+    AppUserDetailsRepo appUserDetailsRepo;
+
+    @BeforeEach
+    public void cleanUp(){
+        boxRepository.deleteAll();
+        appUserLoginRepository.deleteAll();
+        appUserDetailsRepo.deleteAll();
+    }
+
+    String getTokenFor(String username, String password) {
+        return webTestClient.post()
+                .uri("/auth/login")
+                .bodyValue(AppUser.builder()
+                        .username(username)
+                        .password(password)
+                        .build())
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(String.class)
+                .returnResult()
+                .getResponseBody();
+    }
+
+    AppUser createTestUserInLoginRepoAndGet(String id, String name, String password){
+        String hashedPassword = passwordEncoder.encode(password);
+        AppUser testUser = AppUser.builder()
+                .id(id)
+                .username(name)
+                .password(hashedPassword)
+                .build();
+        appUserLoginRepository.save(testUser);
+        return testUser;
+    }
+}
